Move filterLocation func to location-tools.js
